Creando rutas
- Instalando react-router-dom.
- Dividiendo secciones en diferentes pages.
- Creando rutas para cada Pages.
- Agregando links en el navbar para redireccionar a las rutas creadas
