+++ src/App.css
@@ -1,38 +1,42 @@
-.App {
+@media (max-width: 1100px) {
+  .project-grid {
+  grid-template-columns: 1fr 1fr;
+}
+}
+
+@media (max-width: 770px) {
+
+.welcome-description h1{
+  padding: 0;
+}
+
+.project-grid {
+  grid-template-columns: 1fr;
+}
+
+#navbar {
+  justify-content: center;
+}
+
+#navbar a {
   text-align: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
+.social-media__container ul {
   flex-direction: column;
-  align-items: center;
+  line-height: 1.6;
+}
+
+.social-media__container li {
   justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
 }
 
-.App-link {
-  color: #61dafb;
+.social-media__container li:hover {
+  transform: none;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+footer {
+  flex-direction: column;
+  text-align: center;
+}
 }
