+++ src/App.js
@@ -1,24 +1,23 @@
-import logo from './logo.svg';
+import React from 'react';
+import { BrowserRouter as Router, Route } from 'react-router-dom'
+import projectList from './links/projectList';
+import contactList from './links/contactList';
+import Welcome from "./components/pages/welcomeSection/";
+import Grid from './components/pages/grid/';
+import Contact from './components/pages/contactList'
 import './App.css';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <Router>
+      <Route exact path = '/' component = { Welcome } />
+      <Route exact path= '/grid-project'>
+          <Grid projectList = { projectList } />
+      </Route>
+      <Route exact path = '/contact'>
+          <Contact contactList = { contactList } />
+      </Route>
+    </Router>
   );
 }
