--- src/components/organisms/welcome-section/index.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import './welcome.css';
-
-const Welcome = () => {
-
-    return (
-        <section id="welcome-section" class="welcome-section welcome-content">
-            <div class="welcome-description">
-                <h1>¡Hey! mi nombre es Juan Carlos y soy Frontend Developer." <span class="typing">|</span></h1>
-            </div>
-        </section>
-    )
-}
-
-export default Welcome;
