--- src/components/organisms/welcome-section/welcome.css
@@ -1,32 +1,0 @@
-.welcome-section {
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    overflow-x: hidden;
-    text-align: center;
-    background-color: black;
-    justify-content: center;
-    text-align: center;
-  }
-
-  .welcome-description {
-    z-index: 2;
-    background-color: rgba(0,0,0,.3);
-    padding: 0;
-    color: white;
-  }
-
-  .welcome-description {
-    font-size: 30px;
-  }
-
-  .typing {
-    animation: flicker 0.7s infinite;
-  }
-
-  @keyframes flicker {
-    to {
-      color: transparent;
-    }
-  }
