+++ src/components/pages/welcomeSection/welcome.css
@@ -1,0 +1,72 @@
+.welcome-section {
+    height: 100vh;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    overflow-x: hidden;
+    text-align: center;
+    background-color: black;
+    justify-content: center;
+  }
+
+  .welcome-description h1 {
+    /* overflow: hidden; */
+    /* border-right: .15em solid white; */
+    /* white-space: nowrap; */
+    /* letter-spacing: .15em; */
+    font-size: 30px;
+    /* animation: typing 5s steps(41),  */
+    /* animation: fliker .75s step-end infinite; */
+  }
+/*
+  @keyframes typing {
+    from { width: 0 }
+    to { width: 100% }
+  }*/
+
+  @keyframes fliker {
+    from, to { border-color: transparent }
+    50% { border-color: white; }
+  }
+
+  /* .welcome-description {
+    z-index: 2;
+    background-color: rgba(0,0,0,.3);
+    padding: 0;
+    color: white;
+    font-size: 30px;
+
+  } */
+
+  /* h1 {
+    float: left;
+    position: relative;
+  }
+
+  h1 span {
+    position: absolute;
+    right: 0;
+    width: 0;
+    border-left: 1px solid white;
+    background-color: black;
+    animation: typing 6s infinite alternate steps(64);
+  }*/
+
+  .flicker {
+    animation: flicker 0.7s infinite;
+  }
+
+  /* @keyframes typing {
+    from {
+      width: 100%;
+    }
+    to {
+      width: 0;
+    }
+  } */
+
+@keyframes flicker {
+    to {
+      color: transparent;
+    }
+  }
